requirements.txt
Django==4.1.7
djangorestframework==3.14.0
Pillow==9.5.0
python-dotenv==1.0.0
psycopg2-binary
drf-yasg==1.21.5
gunicorn==20.1.0
django-cors-headers==3.13.0
django-crontab==0.7.1
qrcode==7.4.2